PLUGIN: beacon scan compensation moved to ratos.py
