+++ z-probe/beacon.cfg
@@ -25,9 +25,6 @@ contact_max_hotend_temperature: 275
 [bed_mesh]
 mesh_min: 20,30
 
-[beacon_scan_compensator]
-enable_compensation: False
-
 #####
 # BEACON CONFIGURATION
 #####
@@ -51,6 +48,7 @@ variable_beacon_contact_prime_probing: True             # probe for priming with
 variable_beacon_contact_calibration_temp: 150           # nozzle temperature for auto calibration
 variable_beacon_contact_expansion_compensation: True    # enables the nozzle thermal expansion compensation
 variable_beacon_contact_wipe_before_calibrate: True     # enables a nozzle wipe at Y0 before doing the contact calibration
+variable_beacon_scan_compensation_enable: False         # Enables the beacon scan compensation
 variable_beacon_scan_compensation_profile: "Contact"    # The contact profile name for the scan compensation
 
 
@@ -737,8 +735,9 @@ gcode:
 
 	# config
 	{% set default_toolhead = printer["gcode_macro RatOS"].default_toolhead|default(0)|int %}
+	{% set beacon_scan_compensation_enable = true if printer["gcode_macro RatOS"].beacon_scan_compensation_enable|default(false)|lower == 'true' else false %}
 
-	{% if printer["beacon_scan_compensator"] is defined %}
+	{% if beacon_scan_compensation_enable %}
 		# home and abl the printer if needed  
 		_BEACON_HOME_AND_ABL
 
@@ -784,11 +783,12 @@ gcode:
 gcode:
 	# beacon config
 	{% set beacon_scan_compensation_profile = printer["gcode_macro RatOS"].beacon_scan_compensation_profile %}
+	{% set beacon_scan_compensation_enable = true if printer["gcode_macro RatOS"].beacon_scan_compensation_enable|default(false)|lower == 'true' else false %}
 
-	{% if printer["beacon_scan_compensator"] is defined %}
-		{% if printer.configfile.config.beacon_scan_compensator.enable_compensation|default(False)|lower == "true" %}
-			BEACON_APPLY_SCAN_COMPENSATION PROFILE={beacon_scan_compensation_profile}
-		{% endif %}
+	DEBUG_ECHO PREFIX="_BEACON_APPLY_SCAN_COMPENSATION" MSG="beacon_scan_compensation_profile {beacon_scan_compensation_profile}, beacon_scan_compensation_enable {beacon_scan_compensation_enable}"
+
+	{% if beacon_scan_compensation_enable %}
+		BEACON_APPLY_SCAN_COMPENSATION PROFILE={beacon_scan_compensation_profile}
 	{% endif %}
